drop command compat
