+++ lib/vagrant-vbguest/command.rb
@@ -1,12 +1,10 @@
 require 'optparse'
+require Vagrant.source_root.join("plugins/commands/up/start_mixins")
 
 module VagrantVbguest
-
-  module CommandCommons
+  class Command < Vagrant.plugin("2", :command)
+    include VagrantPlugins::CommandUp::StartMixins
     include VagrantVbguest::Helpers::Rebootable
-    def self.included(base)
-      base.extend(ClassMethods)
-    end
 
     # Runs the vbguest installer on the VMs that are represented
     # by this environment.
@@ -72,7 +70,11 @@ module VagrantVbguest
           with_target_vms(vm_name) { |vm| execute_on_vm(vm, options) }
         end
       end
+    end
 
+    # Show description when `vagrant list-commands` is triggered
+    def self.synopsis
+      "plugin: vagrant-vbguest: install VirtualBox Guest Additions to the machine"
     end
 
     protected
@@ -103,8 +105,10 @@ module VagrantVbguest
     end
 
     def check_runable_on(vm)
-      raise NotImplementedError
+      raise Vagrant::Errors::VMNotCreatedError if vm.state.id == :not_created
+      raise Vagrant::Errors::VMInaccessible if vm.state.id == :inaccessible
+      raise Vagrant::Errors::VMNotRunningError if vm.state.id != :running
+      raise VagrantVbguest::NoVirtualBoxMachineError if vm.provider.class != VagrantPlugins::ProviderVirtualBox::Provider
     end
   end
-
 end
