--- lib/vagrant-vbguest/vagrant_compat/vagrant_1_3/command.rb
@@ -1,1 +1,0 @@
-require File.expand_path("../../vagrant_1_2/command", __FILE__)
